requirements.txt
pvporcupine
sounddevice
python-dotenv
wavio
numpy
webrtcvad
setuptools
git+https://github.com/openai/whisper.git
ollama-python
ollama